Refactor inventory components into their own package
