+++ source/core/src/main/com/csse3200/game/components/inventory/PlayerInventoryDisplay.java
@@ -1,4 +1,4 @@
-package com.csse3200.game.components.player;
+package com.csse3200.game.components.inventory;
 
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.graphics.Color;
