+++ source/core/src/main/com/csse3200/game/components/quests/QuestManager.java
@@ -5,7 +5,7 @@ import com.badlogic.gdx.utils.Array;
 import com.csse3200.game.components.Component;
 import com.csse3200.game.components.minigames.MiniGameMedals;
 import com.csse3200.game.components.minigames.MiniGameNames;
-import com.csse3200.game.components.player.PlayerInventoryDisplay;
+import com.csse3200.game.components.inventory.PlayerInventoryDisplay;
 
 import com.csse3200.game.entities.Entity;
 import com.csse3200.game.gamestate.GameState;
