+++ source/core/src/main/com/csse3200/game/components/tasks/ItemProximityTask.java
@@ -1,7 +1,7 @@
 package com.csse3200.game.components.tasks;
 
 import com.badlogic.gdx.scenes.scene2d.Stage;
-import com.csse3200.game.components.player.PlayerInventoryDisplay;
+import com.csse3200.game.components.inventory.PlayerInventoryDisplay;
 import com.csse3200.game.services.DialogueBoxService;
 import com.csse3200.game.entities.Entity;
 import com.csse3200.game.events.EventHandler;
